Rename feature filter-pr-by-build-status → pr-filters
